+++ src/Add-Stats_function.ps1
@@ -1,6 +1,6 @@
 <#
 .SYNOPSIS
-    Add-Stats - Add statistics columns to each record
+    Add-Stats (Alias: astat) - Add statistics columns to each record
 
     For categorical data analysis of time series data.
     Automatically exclude NA, NaN, Null from the specified column.
@@ -9,12 +9,12 @@
         Add-Stats [-v] <col>,<col>,... <params>
 
     Params:
-        [-Sum]
         [-Cnt|-Count]
+        [-Sum]
         [-Mean|-Average]
         [-Max|-Maximum]
         [-Min|-Minimum]
-        [-Sd|-StandardDeviation]
+        [-SD|-StandardDeviation]
         [-All|-AllStats]
         [-Rs] ...Absolute value of deviation
                  from previous record
@@ -126,7 +126,7 @@ function Add-Stats
         [switch] $Rs,
         
         [Parameter( Mandatory=$False)]
-        [Alias('Sd')]
+        [Alias('SD')]
         [switch] $StandardDeviation,
         
         [Parameter( Mandatory=$False)]
@@ -370,3 +370,4 @@ function Add-Stats
         #$obj
     }
 }
+Set-Alias -Name astat -Value Add-Stats