+++ src/percentile_function.ps1
@@ -954,4 +954,3 @@ function percentile {
         }
     }
 }
-