+++ CHANGELOG.md
@@ -10,6 +10,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - NA
 
+## [5.0.3] - 2023-09-23
+
+- Added [Measure-Property] function alias: mprop
+- Added [Add-Stats] function alias: astat
+- Changed [man2] ordinal sort
+
+## [5.0.2] - 2023-09-23
+
+- Fixed [README.md], [Plot-BarChart], [Detect-XrsAnomaly] typo
+- Renamed [Measure-Property] param alias
+- Renamed [Add-Stats] param alias
+
 ## [5.0.1] - 2023-09-22
 
 - Fixed [README.md] typo
@@ -450,7 +462,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [Delete-Field]: src/Delete-Field_function.ps1
 
 
-[unreleased]: https://github.com/btklab/posh-mocks/compare/5.0.1..HEAD
+[unreleased]: https://github.com/btklab/posh-mocks/compare/5.0.3..HEAD
+[5.0.3]: https://github.com/btklab/posh-mocks/releases/tag/5.0.3
+[5.0.2]: https://github.com/btklab/posh-mocks/releases/tag/5.0.2
 [5.0.1]: https://github.com/btklab/posh-mocks/releases/tag/5.0.1
 [5.0.0]: https://github.com/btklab/posh-mocks/releases/tag/5.0.0
 [4.3.0]: https://github.com/btklab/posh-mocks/releases/tag/4.3.0
@@ -495,4 +509,3 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [1.0.0]: https://github.com/btklab/posh-mocks/releases/tag/1.0.0
 [0.2.0]: https://github.com/btklab/posh-mocks/releases/tag/0.2.0
 [0.1.0]: https://github.com/btklab/posh-mocks/releases/tag/0.1.0
-